+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A triangle mesh (.obj), the naive labeling (.txt) and a glTF model are included for all models. The .obj can be downloaded directly from the README.
 - Links to visualize the 3D model with [3dviewer.net](https://3dviewer.net/) ([kovacsv/Online3DViewer](https://github.com/kovacsv/Online3DViewer))
+- `cuboid_tray_with_step` model
 
 ### Changed